Move facet parsing to SolrQueryBaseService class.
