+++ opengever/api/listing.py
@@ -1,4 +1,3 @@
-from collective.elephantvocabulary import wrap_vocabulary
 from DateTime import DateTime
 from DateTime.interfaces import DateTimeError
 from ftw.solr.converters import to_iso8601
@@ -6,11 +5,9 @@ from ftw.solr.interfaces import ISolrSearch
 from ftw.solr.query import escape
 from opengever.api.solr_query_service import SolrQueryBaseService
 from opengever.base.behaviors.translated_title import ITranslatedTitleSupport
-from opengever.base.helpers import display_name
 from opengever.base.interfaces import ISearchSettings
 from opengever.base.solr import OGSolrDocument
 from opengever.base.utils import get_preferred_language_code
-from opengever.globalindex.browser.report import task_type_helper as task_type_value_helper
 from opengever.task.helper import task_type_helper
 from plone import api
 from plone.app.contentlisting.interfaces import IContentListingObject
@@ -23,9 +20,6 @@ from Products.CMFPlone.utils import safe_unicode
 from Products.ZCatalog.Lazy import LazyMap
 from zExceptions import BadRequest
 from zope.component import getUtility
-from zope.component.hooks import getSite
-from zope.globalrequest import getRequest
-from zope.i18n import translate
 from ZPublisher.HTTPRequest import record
 import Missing
 
@@ -199,35 +193,6 @@ FILTERS = {
 }
 
 
-def translate_document_type(document_type):
-    portal = getSite()
-    voc = wrap_vocabulary(
-            'opengever.document.document_types',
-            visible_terms_from_registry='opengever.document.interfaces.'
-                                        'IDocumentType.document_types')(portal)
-    try:
-        term = voc.getTerm(document_type)
-    except LookupError:
-        return document_type
-    else:
-        return term.title
-
-
-def translate_task_type(task_type):
-    return task_type_value_helper(task_type)
-
-
-FACET_TRANSFORMS = {
-    'responsible': display_name,
-    'review_state': lambda state: translate(state, domain='plone',
-                                            context=getRequest()),
-    'document_type': translate_document_type,
-    'task_type': translate_task_type,
-    'checked_out': display_name,
-    'Creator': display_name,
-}
-
-
 def with_active_solr_only(func):
     """Raises an error if solr is not activated
     """
@@ -356,14 +321,6 @@ class Listing(SolrQueryBaseService):
             query=query, filters=filters, start=start, rows=rows, sort=sort,
             fl=field_list, **params)
 
-        # We map the index names back to the field names for the facets
-        facet_counts = {}
-        for field in self.facets:
-            index_name = self.field_name_to_index(field)
-            if index_name is None or index_name not in resp.facets:
-                continue
-            facet_counts[field] = resp.facets[index_name]
-
         items = LazyMap(OGSolrDocument,
                         start * [None] + resp.docs,
                         actual_result_count=resp.num_found,)
@@ -381,16 +338,16 @@ class Listing(SolrQueryBaseService):
         for item in items[start:start + rows]:
             res['items'].append(create_list_item(item, self.columns))
 
-        if facet_counts:
-            for field, facets in facet_counts.items():
-                transform = FACET_TRANSFORMS.get(FIELDS[field][0])
-                for facet, count in facets.items():
-                    facets[facet] = {"count": count}
-                    if transform:
-                        facets[facet]['label'] = transform(facet)
-                    else:
-                        facets[facet]['label'] = facet
-            res['facets'] = facet_counts
+        facet_counts = self.extract_facets_from_response(resp)
+
+        # We map the index names back to the field names for the facets
+        mapped_facet_counts = {}
+        for field in self.facets:
+            index_name = self.field_name_to_index(field)
+            if index_name is None or index_name not in facet_counts:
+                continue
+            mapped_facet_counts[field] = facet_counts[index_name]
+        res['facets'] = mapped_facet_counts
 
         return res
